+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-08-16
+## [Unreleased] - 2022-09-02
 
 ### Added
   
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+* Correctly handle cases of javadoc redirected pages, such as when javadoc generates a
+  redirect from root of javadoc to a module page when Java Platform Module System is used.
+  This includes not inserting a redundant canonical link, and also includes directing search
+  engines to noindex and follow such redirected pages.
 
 ### Dependencies
 * Bump cicirello/pyaction from 4.3.0 to 4.7.1